pyproject.toml
[project]
name = "fiq-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncio>=3.4.3",
    "clickhouse-connect>=0.8.17",
    "flipside>=2.0.8",
    "nbformat>=4.2.0",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.0",
    "web3>=6.20.4",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
